+++ index.html
@@ -235,6 +235,7 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
+  <script src="vendor/pristine/pristine.min.js" type="module" defer></script>
   <script src="js/main.js" type="module" defer></script>
 </body>
 </html>
