+++ js/main.js
@@ -1,7 +1,8 @@
 import { getPhotos } from './data.js';
 import { renderPhotos } from './photos.js';
 import { addGalleryListener } from './gallery.js';
-import './form.js';
+import './open-form.js';
+import './validate-form.js';
 
 const photos = getPhotos();
